+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/enums/SubmissionStatus.kt
@@ -1,0 +1,7 @@
+package uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.enums
+
+enum class SubmissionStatus {
+  NOT_SUBMITTED,
+  SUCCESS,
+  FAILURE,
+}
