+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/fms/FmsSubmissionResult.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models
+package uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.fms
 
 import jakarta.persistence.CascadeType.ALL
 import jakarta.persistence.Column
@@ -12,80 +12,8 @@ import jakarta.persistence.OneToMany
 import jakarta.persistence.OneToOne
 import jakarta.persistence.Table
 import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.enums.FmsOrderSource
-import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.fms.FmsSubmissionStrategyKind
+import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.enums.SubmissionStatus
 import java.util.*
-
-enum class SubmissionStatus {
-  NOT_SUBMITTED,
-  SUCCESS,
-  FAILURE,
-}
-
-@Entity
-@Table(name = "FMS_DEVICE_WEARER_SUBMISSION_RESULT")
-data class FmsDeviceWearerSubmissionResult(
-  @Id
-  @Column(name = "ID", nullable = false, unique = true)
-  val id: UUID = UUID.randomUUID(),
-
-  @Enumerated(EnumType.STRING)
-  @Column(name = "STATUS", nullable = false)
-  val status: SubmissionStatus = SubmissionStatus.NOT_SUBMITTED,
-
-  @Column(name = "PAYLOAD", nullable = false, columnDefinition = "varchar(409600)")
-  val payload: String? = "",
-
-  @Column(name = "DEVICE_WEARER_ID", nullable = false)
-  val deviceWearerId: String? = "",
-
-  @Column(name = "ERROR", nullable = false)
-  val error: String? = "",
-)
-
-@Entity
-@Table(name = "FMS_MONITORING_ORDER_SUBMISSION_RESULT")
-data class FmsMonitoringOrderSubmissionResult(
-  @Id
-  @Column(name = "ID", nullable = false, unique = true)
-  val id: UUID = UUID.randomUUID(),
-
-  @Enumerated(EnumType.STRING)
-  @Column(name = "STATUS", nullable = false)
-  val status: SubmissionStatus = SubmissionStatus.NOT_SUBMITTED,
-
-  @Column(name = "PAYLOAD", nullable = false, columnDefinition = "varchar(409600)")
-  val payload: String? = "",
-
-  @Column(name = "MONITORING_ORDER_ID", nullable = false)
-  val monitoringOrderId: String? = "",
-
-  @Column(name = "ERROR", nullable = false)
-  val error: String? = "",
-)
-
-@Entity
-@Table(name = "FMS_ATTACHMENT_SUBMISSION_RESULT")
-data class FmsAttachmentSubmissionResult(
-  @Id
-  @Column(name = "ID", nullable = false, unique = true)
-  val id: UUID = UUID.randomUUID(),
-
-  @Enumerated(EnumType.STRING)
-  @Column(name = "STATUS", nullable = false)
-  val status: SubmissionStatus = SubmissionStatus.NOT_SUBMITTED,
-
-  @Column(name = "SYS_ID", nullable = false)
-  val sysId: String? = "",
-
-  @Column(name = "FILE_TYPE", nullable = false)
-  val fileType: String? = "",
-
-  @Column(name = "ATTACHMENT_ID", nullable = false)
-  val attachmentId: String? = "",
-
-  @Column(name = "ERROR", nullable = false)
-  val error: String? = "",
-)
 
 @Entity
 @Table(name = "FMS_SUBMISSION_RESULT")
